+++ tests/KeyringController.test.ts
@@ -1,10 +1,15 @@
 import * as ethUtil from 'ethereumjs-util';
+import {
+  recoverPersonalSignature,
+  recoverTypedSignature,
+  recoverTypedSignature_v4,
+  recoverTypedSignatureLegacy,
+} from 'eth-sig-util';
 import { stub } from 'sinon';
 import KeyringController, { Keyring, KeyringConfig } from '../src/keyring/KeyringController';
 import PreferencesController from '../src/user/PreferencesController';
 import ComposableController from '../src/ComposableController';
 
-const sigUtil = require('eth-sig-util');
 const Transaction = require('ethereumjs-tx');
 const mockEncryptor: any = require('./utils/mockEncryptor');
 
@@ -40,10 +45,23 @@ describe('KeyringController', () => {
   });
 
   it('should add new account', async () => {
+    const initialIdentitiesLength = Object.keys(preferences.state.identities).length;
     const currentKeyringMemState = await keyringController.addNewAccount();
     expect(initialState.keyrings).toHaveLength(1);
     expect(initialState.keyrings[0].accounts).not.toBe(currentKeyringMemState.keyrings);
     expect(currentKeyringMemState.keyrings[0].accounts).toHaveLength(2);
+    const identitiesLength = Object.keys(preferences.state.identities).length;
+    expect(identitiesLength).toBeGreaterThan(initialIdentitiesLength);
+  });
+
+  it('should add new account without updating', async () => {
+    const initialIdentitiesLength = Object.keys(preferences.state.identities).length;
+    const currentKeyringMemState = await keyringController.addNewAccountWithoutUpdate();
+    expect(initialState.keyrings).toHaveLength(1);
+    expect(initialState.keyrings[0].accounts).not.toBe(currentKeyringMemState.keyrings);
+    expect(currentKeyringMemState.keyrings[0].accounts).toHaveLength(2);
+    const identitiesLength = Object.keys(preferences.state.identities).length;
+    expect(identitiesLength).toEqual(initialIdentitiesLength);
   });
 
   it('should create new vault and keychain', async () => {
@@ -139,7 +157,7 @@ describe('KeyringController', () => {
     const data = ethUtil.bufferToHex(Buffer.from('Hello from test', 'utf8'));
     const account = initialState.keyrings[0].accounts[0];
     const signature = await keyringController.signPersonalMessage({ data, from: account });
-    const recovered = sigUtil.recoverPersonalSignature({ data, sig: signature });
+    const recovered = recoverPersonalSignature({ data, sig: signature });
     expect(account).toBe(recovered);
   });
 
@@ -158,7 +176,7 @@ describe('KeyringController', () => {
     ];
     const account = initialState.keyrings[0].accounts[0];
     const signature = await keyringController.signTypedMessage({ data: typedMsgParams, from: account }, 'V1');
-    const recovered = sigUtil.recoverTypedSignatureLegacy({ data: typedMsgParams, sig: signature });
+    const recovered = recoverTypedSignatureLegacy({ data: typedMsgParams, sig: signature as string });
     expect(account).toBe(recovered);
   });
 
@@ -199,7 +217,7 @@ describe('KeyringController', () => {
       { data: JSON.stringify(msgParams), from: account },
       'V3',
     );
-    const recovered = sigUtil.recoverTypedSignature({ data: msgParams, sig: signature });
+    const recovered = recoverTypedSignature({ data: msgParams as any, sig: signature as string });
     expect(account).toBe(recovered);
   });
 
@@ -257,7 +275,7 @@ describe('KeyringController', () => {
       { data: JSON.stringify(msgParams), from: account },
       'V4',
     );
-    const recovered = sigUtil.recoverTypedSignature_v4({ data: msgParams, sig: signature });
+    const recovered = recoverTypedSignature_v4({ data: msgParams as any, sig: signature as string });
     expect(account).toBe(recovered);
   });